Add files via upload
This revision includes new independent log files for sub-directories that need to be identified by use of your own domain name if deployed online. Currently, example.com is used, and I've provided a brief description on how to point your sub-directory pages to the root domain in order to make it available (if you so choose) in the ReadMe text document shown here.