Add files via upload
Improved the speed of pages by reducing the number of divs and streamlining the entry point with only one.  Like a modern hull, it's wider at the back closing with two divs, and narrow in the front opening with only one. Am also clarifying the use of the build by adding additional comment tags and improved the display of a new sidebar by writing edits to the w3.css file (which is now also provided in template form w3-template.css).