Add files via upload
To add clarity of purpose to this exercise, I've changed the order of locations described to Nation, State, Jurisdiction to be better able to field information I found troubling. The point of this exercise is to confront the hazards of reporting, which the State requires so by using the power of suggestion (that the nation state the jurisdiction), we're more likely to hold them accountable.